requirements.txt
blinker==1.6.2
boto3==1.28.53
botocore==1.31.53
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
ffmpeg-python==0.2.0
filelock==3.12.4
Flask==2.3.3
future==0.18.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
llvmlite==0.41.0
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.1
numba==0.58.0
numpy==1.25.2
openai-whisper==20230918
python-dateutil==2.8.2
regex==2023.8.8
requests==2.31.0
s3transfer==0.6.2
semantic-version==2.10.0
setuptools-rust==1.7.0
six==1.16.0
sympy==1.12
tiktoken==0.3.3
tomli==2.0.1
torch==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==1.26.16
Werkzeug==2.3.7
